requirements.txt
# ========== 核心框架 ==========
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# ========== 本地大模型推理 ==========
transformers>=4.45.0
torch>=2.2.0
torchvision>=0.17.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
qwen-vl-utils>=0.0.8

# ========== 语音识别 (Whisper) ==========
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# ========== TTS 语音合成 ==========
pyttsx3>=2.90
edge-tts>=6.1.9

# ========== OCR ==========
paddlepaddle>=2.6.0
paddleocr>=2.7.3

# ========== 向量数据库 (记忆系统) ==========
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ========== 屏幕控制 ==========
pyautogui>=0.9.54
pillow>=10.0.0
mss>=9.0.1
opencv-python>=4.9.0

# ========== 系统/进程控制 ==========
psutil>=5.9.0
pywin32>=306
pywinauto>=0.6.8

# ========== Web UI ==========
gradio>=4.40.0

# ========== 工具库 ==========
python-dotenv>=1.0.0
pydantic>=2.7.0
loguru>=0.7.2
aiofiles>=23.2.1
httpx>=0.27.0
